+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ndarray/ShapeInfoProvider.java
@@ -12,21 +12,21 @@ public interface ShapeInfoProvider {
      * @param shape
      * @return
      */
-    Pair<DataBuffer, int[]> createShapeInformation(int[] shape);
+    Pair<DataBuffer, long[]> createShapeInformation(int[] shape);
 
     /**
      * This method creates shapeInformation buffer, based on shape & order being passed in
      * @param shape
      * @return
      */
-    Pair<DataBuffer, int[]> createShapeInformation(int[] shape, char order);
+    Pair<DataBuffer, long[]> createShapeInformation(int[] shape, char order);
 
     /**
      * This method creates shapeInformation buffer, based on detailed shape information being passed in
      * @param shape
      * @return
      */
-    Pair<DataBuffer, int[]> createShapeInformation(int[] shape, int[] stride, long offset, int elementWiseStride, char order);
+    Pair<DataBuffer, long[]> createShapeInformation(int[] shape, int[] stride, long offset, int elementWiseStride, char order);
 
 
     /**
